+++ host/src/FunctionsNetHost/AppLoader/NetHost.cs
@@ -7,18 +7,28 @@ namespace FunctionsNetHost
 {
     internal class NetHost
     {
+        public unsafe struct get_hostfxr_parameters
+        {
+            public nint size;
+
+            // Optional.Path to the application assembly,
+            // If specified, hostfxr is located from this directory if present (For self-contained deployments)
+            public char* assembly_path;
+            public char* dotnet_root;
+        }
+
         [DllImport("nethost", CharSet = CharSet.Auto)]
-        private static extern int get_hostfxr_path(
+        private unsafe static extern int get_hostfxr_path(
         [Out] char[] buffer,
         [In] ref int buffer_size,
-        IntPtr reserved);
+        get_hostfxr_parameters* parameters);
 
-        internal static string GetHostFxrPath()
+        internal unsafe static string GetHostFxrPath(get_hostfxr_parameters* parameters)
         {
             char[] buffer = new char[200];
             int bufferSize = buffer.Length;
 
-            int rc = get_hostfxr_path(buffer, ref bufferSize, IntPtr.Zero);
+            int rc = get_hostfxr_path(buffer, ref bufferSize, parameters);
 
             if (rc != 0)
             {